returning joined calendars
